+++ Server/MirDatabase/RecipeInfo.cs
@@ -77,7 +77,7 @@ namespace Server.MirDatabase
                             switch (data[0].ToLower())
                             {
                                 case "amount":
-                                    Item.Count = uint.Parse(data[1]);
+                                    Item.Count = ushort.Parse(data[1]);
                                     break;
                                 case "chance":
                                     Chance = byte.Parse(data[1]);
@@ -126,9 +126,9 @@ namespace Server.MirDatabase
 
                             UserItem ingredient = Envir.CreateShopItem(info, 0);
 
-                            uint count = 1;
+                            ushort count = 1;
                             if (data.Length >= 2)
-                                uint.TryParse(data[1], out count);
+                                ushort.TryParse(data[1], out count);
 
                             if (data.Length >= 3)
                                 ushort.TryParse(data[2], out ingredient.CurrentDura);
