+++ Shared/ClientPackets.cs
@@ -542,13 +542,13 @@ namespace ClientPackets
 
         public MirGridType Grid;
         public ulong UniqueID;
-        public uint Count;
+        public ushort Count;
 
         protected override void ReadPacket(BinaryReader reader)
         {
             Grid = (MirGridType)reader.ReadByte();
             UniqueID = reader.ReadUInt64();
-            Count = reader.ReadUInt32();
+            Count = reader.ReadUInt16();
         }
         protected override void WritePacket(BinaryWriter writer)
         {
@@ -576,12 +576,12 @@ namespace ClientPackets
         public override short Index { get { return (short)ClientPacketIds.DropItem; } }
 
         public ulong UniqueID;
-        public uint Count;
+        public ushort Count;
 
         protected override void ReadPacket(BinaryReader reader)
         {
             UniqueID = reader.ReadUInt64();
-            Count = reader.ReadUInt32();
+            Count = reader.ReadUInt16();
         }
         protected override void WritePacket(BinaryWriter writer)
         {
@@ -763,13 +763,13 @@ namespace ClientPackets
         public override short Index { get { return (short)ClientPacketIds.BuyItem; } }
 
         public ulong ItemIndex;
-        public uint Count;
+        public ushort Count;
         public PanelType Type;
 
         protected override void ReadPacket(BinaryReader reader)
         {
             ItemIndex = reader.ReadUInt64();
-            Count = reader.ReadUInt32();
+            Count = reader.ReadUInt16();
             Type = (PanelType)reader.ReadByte();
         }
         protected override void WritePacket(BinaryWriter writer)
@@ -784,12 +784,12 @@ namespace ClientPackets
         public override short Index { get { return (short)ClientPacketIds.SellItem; } }
 
         public ulong UniqueID;
-        public uint Count;
+        public ushort Count;
 
         protected override void ReadPacket(BinaryReader reader)
         {
             UniqueID = reader.ReadUInt64();
-            Count = reader.ReadUInt32();
+            Count = reader.ReadUInt16();
         }
         protected override void WritePacket(BinaryWriter writer)
         {
@@ -802,13 +802,13 @@ namespace ClientPackets
         public override short Index { get { return (short)ClientPacketIds.CraftItem; } }
 
         public ulong UniqueID;
-        public uint Count;
+        public ushort Count;
         public int[] Slots;
 
         protected override void ReadPacket(BinaryReader reader)
         {
             UniqueID = reader.ReadUInt64();
-            Count = reader.ReadUInt32();
+            Count = reader.ReadUInt16();
 
             Slots = new int[reader.ReadInt32()];
             for (int i = 0; i < Slots.Length; i++)
